+++ test_conformance/c11_atomics/test_atomics.cpp
@@ -3145,7 +3145,7 @@ public:
     }
 
 private:
-    int _subCaseId;
+    size_t _subCaseId;
     struct TestDefinition _subCase;
 };
